Moved editor code from HDUtils to HDEditorUtils
