+++ ScriptableRenderPipeline/HDRenderPipeline/Editor/HDEditorUtils.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 7e04ef797263c2a47ad576cce5233df9
+timeCreated: 1507109247
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
